requirements.txt
load_dotenv
requests
pandas
numpy
plotly
matplotlib